fix(correctness): file-size overflow, secret persistence, locale parsing, disc dispatch
- MediaFile/MediaScan/DirectoryTree: widen file Size from int to long so files
  over 2 GB don't overflow to a negative/wrong size.
- CredentialManager.RemoveCredentials: SaveStore after a successful delete so the
  secret doesn't reappear on next load.
- Lock: catch UnauthorizedAccessException in IsFileLocked (read-only files), and
  stop killing the process that holds a file during delete-cleanup — it can be an
  in-flight ffmpeg encode; log and skip instead.
- DoubleConverter: parse with InvariantCulture (JSON is always period-decimal, so
  '7.5' became 75 under a comma-decimal host locale) and drop the bogus /1,000,000.
- MusicBrainzTrack: bind the raw JSON token so the vinyl/box-set (A1, 1-05) number
  fallback is actually reachable instead of dead code behind a never-throwing cast.
- DiscRipJob: dispatch a VideoEncodeJob per ripped title, not once per season
  folder, so episodes 2..N of a disc are no longer dropped.
- KnownForDto: fix NumberOfItems lifted int?+int? that returned null (losing the
  TV-side count) whenever Movie was null.

+++ src/NoMercy.NmSystem/Dto/MediaScan.cs
@@ -424,7 +424,7 @@ public class MediaScan : IDisposable, IAsyncDisposable
                         Name = Path.GetFileName(file),
                         Path = file,
                         Extension = extension,
-                        Size = (int)_driver.GetFileSize(file),
+                        Size = _driver.GetFileSize(file),
                         Created = _driver.GetCreationTimeUtc(file).ToLocalTime(),
                         Modified = _driver.GetLastWriteTimeUtc(file).ToLocalTime(),
                         Accessed = _driver.GetLastAccessTimeUtc(file).ToLocalTime(),
